Perform DE of multiple samples in a screen vs control

Perform DE of multiple samples in a screen vs control

compute_multi_de(
  data = NULL,
  treatment_samples = NULL,
  control_samples = NULL,
  method = "edgeR",
  num_cores = 2,
  batch = 1,
  k = 2,
  spikes = NULL
)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column"

treatment_samples

Value in the column "combined_id" representing replicates of treatment samples in the data

control_samples

Value in the column "combined_id" representing replicates of control samples in the data

method

One of "Seurat_wilcox", "DESeq2", "edgeR", "RUVg", "RUVs", "RUVr", "limma_voom"

num_cores

Number of cores

batch

Either empty, a single value, or a vector corresponding to the number of samples

k

Parameter k for RUVSeq methods, check RUVSeq tutorial

spikes

List of genes to use as spike controls

Value

List of DE counts vs control

Examples

file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)[50:120]
control_samples <- "DMSO_0"
treatment_samples <- mac$combined_id[!grepl("DMSO", mac$combined_id)]
mac <- compute_multi_de(mac, treatment_samples, control_samples, num_cores = 1, method = "edgeR")