Perform enrichR-style analysis on a screen

Perform enrichR-style analysis on a screen

compute_multi_enrichr(
  data,
  genesets = NULL,
  species = NULL,
  direction = "both",
  p_value_cutoff = 0.01,
  n_distinct = 10
)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column".

genesets

Named list of genes

species

One of "human", "mouse", "fly", "yeast", "worm" or "fish"

direction

Direction of the differentially expressed genes, one of "up", "down", "both" (default).

p_value_cutoff

Cutoff for adjusted p-value (column p_value_adj), default 0.01

n_distinct

Minimum number of genes in a geneset, default 5

Value

A tidyseurat object with appended pathway_enrichment dataframe in slot tools

Examples

file_path <- system.file("extdata", "PMMSq033/PMMSq033.rds", package = "macpie")
mac <- readRDS(file_path)
file_path <- system.file("extdata", "PMMSq033/pathways.Rds", package = "macpie")
genesets <- readRDS(file_path)
compute_multi_enrichr(mac, genesets = genesets)
#> Warning: data layer is not found and counts layer is used
#> # A Seurat-tibble abstraction: 308 × 21
#> # Features=21381 | Cells=308 | Active assay=RNA | Assays=RNA
#>    .cell     orig.ident nCount_RNA nFeature_RNA percent.mt percent.ribo Plate_ID
#>    <chr>     <fct>           <dbl>        <int>      <dbl>        <dbl> <chr>   
#>  1 AACAGGCA… PMMSq033         8001         3094      0.962         8.19 PMMSq033
#>  2 AACCAGCC… PMMSq033        69701        12649      0.211         2.56 PMMSq033
#>  3 AACCAGTT… PMMSq033        52404        11309      0.204         2.77 PMMSq033
#>  4 AACCGGCG… PMMSq033        71499        12849      0.203         2.32 PMMSq033
#>  5 AACCTAGT… PMMSq033        38469         9689      0.294         2.74 PMMSq033
#>  6 AACTCTAC… PMMSq033        67371        12665      0.194         2.12 PMMSq033
#>  7 AACTGTGT… PMMSq033        53476        11264      0.258         2.15 PMMSq033
#>  8 AAGATGTC… PMMSq033        43015        10336      0.274         2.53 PMMSq033
#>  9 AAGCATAT… PMMSq033        64145        12023      0.214         2.40 PMMSq033
#> 10 AAGCTCAC… PMMSq033        69879        12851      0.205         2.33 PMMSq033
#> # ℹ 298 more rows
#> # ℹ 14 more variables: Well_ID <chr>, Row <chr>, Column <int>, Species <chr>,
#> #   Cell_type <chr>, Model_type <chr>, Time <int>, Unit <chr>,
#> #   Treatment_1 <chr>, Concentration_1 <dbl>, Unit_1 <chr>, Sample_type <chr>,
#> #   Project <chr>, combined_id <chr>