Retrieve normalised counts of MAC-seq data

This function retrieves counts from a number of methods that are available for normalisation, with the default of limma-voomLmFit.

Usage

compute_normalised_counts(
  data = NULL,
  method = NULL,
  batch = NULL,
  k = NULL,
  spikes = NULL,
  max_counts = NULL,
  num_cores = NULL
)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column"

method

One of "raw", "logNorm", "cpm", "clr", "SCT", "DESeq2", "edgeR", "RUVg", "RUVs", "RUVr", "limma_voom", "limma_trend", "zinb"

batch

Either empty, a single value, or a vector corresponding to the number of samples

k

Parameter k for RUVSeq and zinb methods

spikes

List of genes to use as spike controls

max_counts

Maximum count for a gene across all samples

num_cores

Number of cores

Value

Data frame of normalised counts

Examples

data(mini_mac)
compute_normalised_counts(mini_mac)
#>                 AACAGGCAAT AACCAGCCAG AACCAGTTGA AACCGGCGTA AACCTAGTCC
#> NAMPT                 13.7      14.26       14.4      13.91       13.6
#> ENSG00000278869       12.1       9.87       10.2       9.66       10.8
#> CABP7-DT              12.1       9.87       10.2       9.66       10.8
#>                 AACTCTACAC AACTGTGTCA AAGATGTCCA AAGCATATGG AAGCTCACCT
#> NAMPT                 14.0       13.5       14.1      13.84       14.5
#> ENSG00000278869        9.8       10.3       10.6       9.94        9.9
#> CABP7-DT               9.8       10.3       10.6       9.94        9.9
#>                 AAGGCATGCG AAGTTCCTTG AATACCGGTA AATCCATCTG AATCGTCCGC
#> NAMPT                 14.1      14.24       12.7      14.39       14.2
#> ENSG00000278869       10.4       9.99       10.3       9.87       10.3
#> CABP7-DT              10.4       9.99       10.3       9.87       10.3
#>                 AATGACCTTC AATGAGAGCA AATGTCAGTG AATTAGGCCG AATTGCGATG
#> NAMPT                 14.1       13.8       14.2       14.8       14.0
#> ENSG00000278869       10.2       10.4       10.1       10.3       10.3
#> CABP7-DT              10.2       10.4       10.1       10.3       10.3
#>                 ACAACAGTCG ACAACCATAC ACAACGGAGC ACAAGCGCGA ACACAATCTC
#> NAMPT                 14.2       13.7       13.3       13.9       14.2
#> ENSG00000278869       10.3       11.8       10.5       10.2        9.9
#> CABP7-DT              10.3       10.2       10.5       10.2        9.9
#>                 ACACAGTGAA ACACCGAATT ACACGCAGTA ACACGGTCCT ACACTTGCTG
#> NAMPT                 13.8       10.3       13.3       14.4       12.6
#> ENSG00000278869       11.0       10.3       10.2       10.0       10.3
#> CABP7-DT              11.0       10.3       10.2       10.0       10.3
#>                 ACAGTGCCAA ACATGTGTGC ACCAGGACCA ACCATAACAC ACCGAACCGT
#> NAMPT                 12.6      14.46       14.8       13.9       13.7
#> ENSG00000278869       11.0       9.93       10.1       10.2       10.0
#> CABP7-DT              11.0       9.93       10.1       10.2       10.0
#>                 ACCGAGAGTC ACCGGTACAG ACCGTACTTC ACCTCCGACA ACCTCTCTCC
#> NAMPT                 13.6       13.3       13.6       13.8       13.7
#> ENSG00000278869       10.4       10.5       10.4       10.3       10.2
#> CABP7-DT              10.4       10.5       10.4       10.3       10.2
#>                 ACCTGTCCGA ACGAATGACA ACGCTGGATA ACGTAGGCAC ACGTGCTGAT
#> NAMPT                 13.9      13.38       14.1       13.6       14.1
#> ENSG00000278869       10.4       9.92       10.4       10.4       10.4
#> CABP7-DT              10.4       9.92       10.4       10.4       10.4
#>                 ACTCCAAGCC ACTGGCGCAT ACTGGCTTCC ACTTAACTGC ACTTCATCAC
#> NAMPT                 14.3       14.2       13.2      14.52       13.9
#> ENSG00000278869       10.1       10.1       10.9       9.77       10.2
#> CABP7-DT              10.1       10.1       10.9       9.77       10.2
#>                 ACTTCGTTGA ACTTCTCCTG ACTTGAGGAA ACTTGTAAGG AGAACCACGG
#> NAMPT                 14.0       12.9      13.44       12.8       14.2
#> ENSG00000278869       10.1       10.6       9.98       10.4       10.5
#> CABP7-DT              10.1       10.6       9.98       10.4       10.5
#>                 AGAAGCAATC AGACCGTTAT AGACTAGCAT AGAGATGCAG AGAGCTTACA
#> NAMPT                 14.2      14.21       13.9       14.0       14.3
#> ENSG00000278869       10.1       9.96       10.8       10.3       10.0
#> CABP7-DT              10.1       9.96       10.8       10.3       10.0
#>                 AGAGTGTAAC AGAGTTCTGC AGATAGTGCT AGCAATGCGC AGCATGTCAT
#> NAMPT                 13.7       14.5      13.41       14.0       14.2
#> ENSG00000278869       10.2       10.6       9.95       10.0       10.3
#> CABP7-DT              10.2       10.6       9.95       10.0       10.3
#>                 AGCCACTAGC AGCCAGAATA AGCCAGCTCT AGCGATAACG AGCTATTCCA
#> NAMPT                 14.4      14.05      13.90       13.1      14.36
#> ENSG00000278869       10.1       9.96       9.99       10.3       9.96
#> CABP7-DT              10.1       9.96       9.99       10.3       9.96
#>                 AGGCGTCTGT AGGTCACCAA AGTAACTCAC AGTAAGCGTT AGTCCACGTA
#> NAMPT                 14.6       13.7       13.9       14.1       14.1
#> ENSG00000278869       10.2       10.2       10.2       10.4       12.3
#> CABP7-DT              10.2       10.2       10.2       10.4       10.0
#>                 AGTCTCGGCA AGTCTGTACG AGTGCAATGT ATAACGCCTC ATAAGAGGTC
#> NAMPT                 14.5       13.4       13.8      14.11       14.2
#> ENSG00000278869       10.4       10.2       10.4       9.87       10.5
#> CABP7-DT              10.4       10.2       10.4       9.87       10.5
#>                 ATAAGGTGCA ATACACGACA ATACCTCCGG ATAGCAGTGC ATAGGCCATT
#> NAMPT                 13.6       14.2      13.38       12.6       13.9
#> ENSG00000278869       10.1       10.2       9.92       10.3       10.0
#> CABP7-DT              10.1       10.2       9.92       10.3       10.0
#>                 ATATCCGCAT ATCAGCACTT ATCAGCGAGG ATCCAATACG ATCCGCTGTG
#> NAMPT                 13.2       14.3       14.3      13.84       13.5
#> ENSG00000278869       10.4       10.4       10.4       9.93       10.3
#> CABP7-DT              10.4       10.4       10.4       9.93       10.3
#>                 ATCCGTCCAT ATCGACGGCT ATCGCGATTA ATCGGTAGGC ATCTAAGGAG
#> NAMPT                13.66       14.3       13.7       13.8       14.4
#> ENSG00000278869       9.96       10.4       10.0       10.1        9.9
#> CABP7-DT              9.96       10.4       10.0       10.1        9.9
#>                 ATGACGGTAA ATGACTCAGT ATGCGGACTG ATGGTCTTAG ATGTGGAAGC
#> NAMPT                 14.5      13.84      14.71       13.3       13.9
#> ENSG00000278869       10.2       9.93       9.85       10.5       10.4
#> CABP7-DT              10.2       9.93       9.85       10.5       10.4
#>                 ATTATCGGAC ATTCGGAACA CAACAATCCA CAAGAAGCAT CAAGATGAGG
#> NAMPT                 14.4       12.8       14.1       13.8      13.76
#> ENSG00000278869       10.3       10.5       10.2       10.3       9.85
#> CABP7-DT              10.3       10.5       10.2       10.3       9.85
#>                 CAAGCCAACG CAAGTGGATC CACAGTTCAT CACGAGTCTG CACGCTCCAA
#> NAMPT                 12.8      14.13       14.0       14.6       14.1
#> ENSG00000278869       11.2       9.89       10.3       10.1       10.4
#> CABP7-DT              11.2       9.89       10.3       10.1       10.4
#>                 CACTGAGCAC CAGATCAATG CAGTGCTCTT CAGTTAAGCA CATAGCTATC
#> NAMPT                 14.4       13.8      14.23       13.8       13.9
#> ENSG00000278869       10.1       10.4       9.98       10.4       10.2
#> CABP7-DT              10.1       10.4       9.98       10.4       10.2
#>                 CATCACCACC CATGTACGCC CATTACACTG CATTCGACGA CCAACTATGG
#> NAMPT                 14.8       13.9       14.1      13.90       12.9
#> ENSG00000278869       10.5       10.0       10.4       9.99       10.6
#> CABP7-DT              10.5       10.0       10.4       9.99       10.6
#>                 CCAAGGAGTT CCAATTGTTC CCACAAGTGC CCAGCTTAGT CCATAACTTG
#> NAMPT                14.45       14.4      14.46       14.1      13.62
#> ENSG00000278869       9.81       10.3       9.94       10.9       9.72
#> CABP7-DT              9.81       10.3       9.94       10.9       9.72
#>                 CCATAGATCA CCGAACAAGC CCGAAGACCT CCGAATAGTG CCGACTTCTC
#> NAMPT                 14.2       13.6       14.3       14.4       14.7
#> ENSG00000278869        9.8       12.1       10.6       10.0       10.3
#> CABP7-DT               9.8       10.5       10.6       10.0       10.3
#>                 CCGATCCACT CCGATGATAC CCGCGTTATG CCGCTAGCTT CCGGAGTATC
#> NAMPT                 13.6       14.0       13.9       14.1       14.8
#> ENSG00000278869       12.0       10.6       10.2       10.2       10.1
#> CABP7-DT              10.4       10.6       10.2       10.2       10.1
#>                 CCGGCCAATT CCGGTCTCTA CCGTACGATG CCGTCAGAAC CCTAGACACG
#> NAMPT                 14.0       14.6       13.4      14.17       13.9
#> ENSG00000278869       10.3       10.1       10.6       9.77       10.2
#> CABP7-DT              10.3       10.1       10.6       9.77       10.2
#>                 CCTAGTTGAG CCTATTCTGT CCTCAACCGA CCTCCATAAG CCTGATGCCA
#> NAMPT                 14.4       14.8       14.6       14.1      13.95
#> ENSG00000278869       10.2       10.6       10.2       10.2       9.86
#> CABP7-DT              10.2       10.6       10.2       10.2       9.86
#>                 CCTGCAATAC CCTTGTATTC CGAGATCTCT CGAGGAACAA CGATAACCGC
#> NAMPT                 13.5       14.0       13.4       14.7       14.2
#> ENSG00000278869       10.3       10.1       10.6       11.7       10.3
#> CABP7-DT              10.3       10.1       10.6       10.1       10.3
#>                 CGATCCTGTG CGCCAACCAT CGCCAGTGTT CGCGGATTCA CGCTTCTTGG
#> NAMPT                14.07       13.8       14.5       13.7       14.5
#> ENSG00000278869       9.98       11.4       10.1       10.0       10.3
#> CABP7-DT              9.98       11.4       10.1       10.0       10.3
#>                 CGGAATACAC CGGAGATTGG CGGAGCTCAA CGGATCGGTA CGGATTCTAG
#> NAMPT                 14.1       14.5       14.0      14.50       13.1
#> ENSG00000278869       10.4       10.4       10.3       9.86       10.3
#> CABP7-DT              10.4       10.4       10.3       9.86       10.3
#>                 CGGCAACTTA CGGCTCATCA CGGTCGTATT CGGTGACATC CGTAACGGAT
#> NAMPT                 12.8       14.1       14.2       12.9       14.3
#> ENSG00000278869       10.0       10.4       10.1       10.6       10.1
#> CABP7-DT              10.0       10.4       10.1       10.6       10.1
#>                 CGTAAGATTC CGTACTGTAA CGTAGAAGAC CGTCCTAGGA CGTCGGCAAT
#> NAMPT                 13.6       13.6       14.5      13.89       13.5
#> ENSG00000278869       10.2       10.1       10.4       9.98       10.3
#> CABP7-DT              10.2       10.1       10.4       9.98       10.3
#>                 CGTGAGTTAT CGTGTCAAGC CTAACTTCAG CTAATAGCGT CTACACCAGG
#> NAMPT                 14.2       12.9       14.4       13.9      14.20
#> ENSG00000278869       10.1       10.5       10.0       12.0       9.96
#> CABP7-DT              10.1       10.5       10.0       10.4       9.96
#>                 CTAGCACAAT CTATGAACGG CTCAAGGACC CTCACCTGTC CTCCTATTGT
#> NAMPT                13.97      14.28       14.0       14.4      13.52
#> ENSG00000278869       9.89       9.89       10.3       10.0       9.82
#> CABP7-DT              9.89       9.89       10.3       10.0       9.82
#>                 CTCGCAACGT CTGGATTGAC CTGTCTGTGT CTTATGTTGC CTTCATATCG
#> NAMPT                 13.9       14.3       13.6       14.1       13.9
#> ENSG00000278869       10.0       10.2       10.4       10.2        9.8
#> CABP7-DT              10.0       10.2       10.4       10.2        9.8
#>                 CTTGCTGACG CTTGGATCTT GAAGAGTTCT GAAGGATTAG GAATCGAGCC
#> NAMPT                 13.5       14.1       14.6       13.9       13.2
#> ENSG00000278869       10.7       10.0       10.2       10.2       10.4
#> CABP7-DT              10.7       10.0       10.2       10.2       10.4
#>                 GAATCTTCTC GAATTACGGC GACCATCTAA GACGACCACA GAGAACGAAG
#> NAMPT                 13.2       14.3      14.17       14.2       14.3
#> ENSG00000278869       10.0       10.4       9.92       10.5       11.7
#> CABP7-DT              10.0       10.4       9.92       10.5       10.1
#>                 GAGACAAGGC GAGACATCTT GAGCGAGTCA GAGTAGACCA GATACGCTTA
#> NAMPT                 14.2       13.3       12.8       13.9       12.7
#> ENSG00000278869       11.9       10.1       10.5       10.0       11.1
#> CABP7-DT              10.3       10.1       10.5       10.0       11.1
#>                 GATAGACTGT GATAGAGGCG GATAGGTCAA GATATCAGGA GATCTCATTC
#> NAMPT                 13.0       13.5       14.2       13.8       13.6
#> ENSG00000278869       10.6       10.7       10.1       10.3       10.1
#> CABP7-DT              10.6       10.7       10.1       10.3       10.1
#>                 GATCTGGTCG GATGAGTGAC GATGGATACA GATGTGACAG GATTGCACGC
#> NAMPT                 13.4      14.23       14.7       14.0       10.9
#> ENSG00000278869       10.3       9.98       10.2       10.1       10.9
#> CABP7-DT              10.3       9.98       10.2       10.1       10.9
#>                 GCACACTATA GCACTGCGTT GCACTTAATC GCAGGAGATG GCAGTACTGG
#> NAMPT                 14.1       13.5      13.88       13.9       14.5
#> ENSG00000278869       10.4       10.3       9.97       10.4       10.0
#> CABP7-DT              10.4       10.3       9.97       10.4       10.0
#>                 GCATATGAGT GCATCCGATC GCCAAGTACA GCCACGATTC GCCATAGGTT
#> NAMPT                 14.5       14.3       13.4       14.4       13.7
#> ENSG00000278869       10.2       10.0       10.3       10.0       10.2
#> CABP7-DT              10.2       10.0       10.3       10.0       10.2
#>                 GCCATATCGA GCCGTCAATA GCCTGGACAT GCGTAATTAC GCTATTATCC
#> NAMPT                14.14       13.8      14.09       13.8      13.49
#> ENSG00000278869       9.89       10.3      11.43       10.4       9.79
#> CABP7-DT              9.89       10.3       9.84       10.4       9.79
#>                 GCTCAGTAAT GCTGCTTATA GGAATAAGCA GGACGATGCT GGCATCGTGA
#> NAMPT                 13.6       13.8       14.3       12.4       13.6
#> ENSG00000278869       10.8       10.1       10.4       10.1       10.4
#> CABP7-DT              10.8       10.1       10.4       10.1       10.4
#>                 GGCATTATTG GGCCGAGATT GGCGCTATAA GGCGTTAAGT GGCTATTGAT
#> NAMPT                 14.0       13.6      14.16       13.7       13.2
#> ENSG00000278869       10.1       10.2      11.50       10.2       10.0
#> CABP7-DT              10.1       10.2       9.91       10.2       10.0
#>                 GGCTGCTACT GGTAATGTGT GGTGTTCACC GGTTCACTGG GTAACCAGTA
#> NAMPT                 13.8       14.3       14.5       14.0       13.6
#> ENSG00000278869       10.3       10.0       10.2       10.6       10.2
#> CABP7-DT              10.3       10.0       10.2       10.6       10.2
#>                 GTAACCTTGG GTAAGAACCT GTAAGGCTCC GTAATCCACG GTATTGTGGA
#> NAMPT                 13.7       13.8       13.7       13.5       14.2
#> ENSG00000278869       10.0       10.1       10.2       10.0       10.1
#> CABP7-DT              10.0       10.1       10.2       10.0       10.1
#>                 GTCCGCATCA GTCCTTCGGT GTCGCTCTCT GTCGGTGACA GTCTCGAGTG
#> NAMPT                 14.1       15.1       14.0      13.31       13.6
#> ENSG00000278869       10.4       10.2       10.3       9.85       10.5
#> CABP7-DT              10.4       10.2       10.3       9.85       10.5
#>                 GTCTCTTAAG GTCTTCCGAG GTGACTATAC GTGGTTAATG GTGTGCCTGT
#> NAMPT                 14.3       13.7      13.94       14.2       13.7
#> ENSG00000278869       10.4       10.5       9.86       10.3       10.3
#> CABP7-DT              10.4       10.5       9.86       10.3       10.3
#>                 GTGTGTGTCC GTTCATTGCC GTTCCGGTGA GTTCGTCGAA GTTGAATTGG
#> NAMPT                 14.3       14.2       13.3      13.87       13.6
#> ENSG00000278869       10.2       10.1       10.5       9.96       10.4
#> CABP7-DT              10.2       10.1       10.5       9.96       10.4
#>                 GTTGATCCGC GTTGTATGCT TAACCGTAGC TAACGTCGAT TAAGGTACGG
#> NAMPT                12.70      14.04       14.3       14.1      14.60
#> ENSG00000278869       9.89       9.96       10.1       10.2       9.84
#> CABP7-DT              9.89       9.96       10.1       10.2       9.84
#>                 TACTACCGCC TAGCGCCAAC TAGGTTGCAA TAGTAGTCTC TAGTCCGCTG
#> NAMPT                14.57       13.5       14.0      14.28       13.6
#> ENSG00000278869       9.93       10.7       10.3       9.89       10.5
#> CABP7-DT              9.93       10.7       10.3       9.89       10.5
#>                 TAGTGGAACT TATCATGCAG TATCGTTACG TCAAGTGCAG TCACAGATAC
#> NAMPT                 12.6       13.3       14.0       13.8       14.6
#> ENSG00000278869       10.3       10.2       10.1       10.3       10.0
#> CABP7-DT              10.3       10.2       10.1       10.3       10.0
#>                 TCACCGCCTA TCACGCCACT TCACGTTGGC TCATTGTCCA TCCACACTAG
#> NAMPT                 14.0       14.0       13.5       14.6       13.0
#> ENSG00000278869       10.3       10.1       10.3       10.1       10.2
#> CABP7-DT              10.3       10.1       10.3       10.1       10.2
#>                 TCCACGGTCA TCCACTCGCT TCCGACTAAC TCCGTTATCT TCCTAAGAGA
#> NAMPT                 14.6       12.3       14.7       14.1       14.3
#> ENSG00000278869       10.1       10.8       10.2       10.2       10.4
#> CABP7-DT              10.1       10.8       10.2       10.2       10.4
#>                 TCCTCTAGTA TCGAAGCATT TCGAGAGAGC TCGCACTTGA TCGCCTACTG
#> NAMPT                 13.6       13.8       12.0      14.83       12.7
#> ENSG00000278869       10.8       10.6       10.4       9.88       10.4
#> CABP7-DT              10.8       10.6       10.4       9.88       10.4
#>                 TCGCGTAGCA TCGGCGTTAA TCTACATCCG TCTCTCCTAT TGAAGAAGGT
#> NAMPT                 13.5       14.4       14.6      14.93       14.9
#> ENSG00000278869       10.4       10.0       10.5       9.97       10.3
#> CABP7-DT              10.4       10.0       10.5       9.97       10.3
#>                 TGAGCGTTCC TGAGTACGTA TGGAATGGAG TGTCATTCGC TGTGCTTCAG
#> NAMPT                 14.5       13.3       13.7       14.6      14.20
#> ENSG00000278869       10.1       10.5       10.2       10.5      11.53
#> CABP7-DT              10.1       10.5       10.2       10.5       9.95
#>                 TGTTCAGGAT TTACACACGT TTACTGTGAC TTATGCCGCG TTCACGGAAG
#> NAMPT                 13.6      14.57       14.5       14.5       14.3
#> ENSG00000278869       10.2       9.71       10.6       10.1       10.4
#> CABP7-DT              10.2       9.71       10.6       10.1       10.4
#>                 TTCGAGTGAT TTCTGTACCT TTGGTAACAG
#> NAMPT                 14.6       14.2      14.30
#> ENSG00000278869       10.2       10.7       9.91
#> CABP7-DT              10.2       10.7       9.91
#>  [ reached getOption("max.print") -- omitted 497 rows ]