Plot UMAP dimensionality reduction on DE genes

Plot UMAP dimensionality reduction on DE genes

Usage

plot_de_umap(data = NULL, color_by = NULL, label = NULL, max_overlaps = NULL)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column"

color_by

A string specifying which column in data will be used to color the samples.

label

A string specifying which column in data will be used to label a sample.

max_overlaps

Maximum number of overlaps for ggrepel

Value

ggplot object