Plot MDS dimensionality reduction

This function uses limma's plot MDS to visualise groupoing of the data points. The function is used with gene.selection = "common" to increase the speed.

Usage

plot_mds(
  data = NULL,
  group_by = NULL,
  label = NULL,
  max_overlaps = NULL,
  n_labels = 50
)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column"

group_by

A string specifying which column in data will be used to color the samples.

label

A string specifying which column in data will be used to label a sample.

max_overlaps

Maximum number of overlaps for ggrepel

n_labels

An integer specifying number of labels to show, based on PC1 and PC2 extremeness. Default set to 50

Value

ggplot object

Examples

data("mini_mac")
p <- plot_mds(mini_mac, group_by = "Sample_type", label = "Sample_type", 
n_labels = 10, max_overlaps = 100)
#> tidyseurat says: Key columns are missing. A data frame is returned for independent data analysis.
#> tidyseurat says: Key columns are missing. A data frame is returned for independent data analysis.