Check and clean the metadata file generated from findmetadata function

It checks the column completeness and removes special characters such as spaces and comma except from Treatment_1

It replaces special characters such as spaces and comma with underscore

It generates a quick summary of the different treatment and sample groups in the metadata

Usage

validate_metadata(metadata)

Arguments

metadata

this is the output from findmetadata function

Value

A list containing:

cleaned_metadata

The cleaned metadata data frame after validation and modifications.

summary_table

A summary table grouped by Plate_ID, showing the count of unique values in selected columns.

Details

This function validates a metadata dataset, cleans specific fields, checks for issues, and generates a summary table grouped by Plate_ID.

Author

Mark Li

Examples

if (FALSE) { # \dontrun{
# Example CSV file
file_path <- system.file("/extdata/PMMSq033_metadata.csv", package = "macpie")
metadata <- read_metadata(file_path)
metadt_qc <- validate_metadata(metadata)
} # }