Retrieve normalised counts of MAC-seq data

This function retrieves counts from a number of methods that are available for normalisation, with the default of limma-voomLmFit.

Usage

compute_normalised_counts(
  data = NULL,
  method = NULL,
  batch = NULL,
  k = NULL,
  spikes = NULL,
  max_counts = NULL
)

Arguments

data

A tidyseurat object merged with metadata. Must contain columns "Well_ID", "Row", "Column"

method

One of "raw", "logNorm", "cpm", "clr", "SCT", "DESeq2", "edgeR", "RUVg", "RUVs", "RUVr", "limma_voom"

batch

Either empty, a single value, or a vector corresponding to the number of samples

k

Parameter k for RUVSeq methods, check RUVSeq tutorial

spikes

List of genes to use as spike controls

max_counts

Maximum count for a gene across all samples

Value

Data frame of normalised counts

Examples

data(mini_mac)
compute_normalised_counts(mini_mac)
#>                 AACAGGCAAT AACCAGCCAG AACCAGTTGA AACCGGCGTA AACCTAGTCC
#> NAMPT             13.66997  14.260025   14.44157  13.909156   13.56072
#> ENSG00000278869   12.08500   9.867707   10.19365   9.661228   10.75337
#> CABP7-DT          12.08500   9.867707   10.19365   9.661228   10.75337
#>                 AACTCTACAC AACTGTGTCA AAGATGTCCA AAGCATATGG AAGCTCACCT
#> NAMPT            14.045187   13.47592   14.06636  13.843252  14.546433
#> ENSG00000278869   9.797259   10.30599   10.60693   9.936361   9.902577
#> CABP7-DT          9.797259   10.30599   10.60693   9.936361   9.902577
#>                 AAGGCATGCG AAGTTCCTTG AATACCGGTA AATCCATCTG AATCGTCCGC
#> NAMPT             14.09024  14.236369   12.65557  14.390264   14.19258
#> ENSG00000278869   10.38980   9.988441   10.33364   9.866702   10.28569
#> CABP7-DT          10.38980   9.988441   10.33364   9.866702   10.28569
#>                 AATGACCTTC AATGAGAGCA AATGTCAGTG AATTAGGCCG AATTGCGATG
#> NAMPT             14.13132   13.84006   14.23717   14.78534   14.04032
#> ENSG00000278869   10.22443   10.38063   10.14970   10.26178   10.33988
#> CABP7-DT          10.22443   10.38063   10.14970   10.26178   10.33988
#>                 ACAACAGTCG ACAACCATAC ACAACGGAGC ACAAGCGCGA ACACAATCTC
#> NAMPT             14.22746   13.67637   13.29803   13.91466  14.152842
#> ENSG00000278869   10.32057   11.80190   10.49068   10.21422   9.904915
#> CABP7-DT          10.32057   10.21694   10.49068   10.21422   9.904915
#>                 ACACAGTGAA ACACCGAATT ACACGCAGTA ACACGGTCCT ACACTTGCTG
#> NAMPT             13.77927   10.32576   13.33117  14.389019   12.62004
#> ENSG00000278869   10.97192   10.32576   10.16124   9.996701   10.29811
#> CABP7-DT          10.97192   10.32576   10.16124   9.996701   10.29811
#>                 ACAGTGCCAA ACATGTGTGC ACCAGGACCA ACCATAACAC ACCGAACCGT
#> NAMPT             12.62198  14.458157   14.78237   13.93570   13.72066
#> ENSG00000278869   11.03701   9.934595   10.13852   10.23526   10.02022
#> CABP7-DT          11.03701   9.934595   10.13852   10.23526   10.02022
#>                 ACCGAGAGTC ACCGGTACAG ACCGTACTTC ACCTCCGACA ACCTCTCTCC
#> NAMPT             13.58220   13.33547   13.61461   13.80380   13.68839
#> ENSG00000278869   10.41227   10.52812   10.44468   10.34436   10.22896
#> CABP7-DT          10.41227   10.52812   10.44468   10.34436   10.22896
#>                 ACCTGTCCGA ACGAATGACA ACGCTGGATA ACGTAGGCAC ACGTGCTGAT
#> NAMPT             13.90623  13.378382   14.08211   13.58088   14.05796
#> ENSG00000278869   10.44680   9.918951   10.38167   10.41096   10.35752
#> CABP7-DT          10.44680   9.918951   10.38167   10.41096   10.35752
#>                 ACTCCAAGCC ACTGGCGCAT ACTGGCTTCC ACTTAACTGC ACTTCATCAC
#> NAMPT             14.32798   14.19152   13.23157  14.524254   13.89201
#> ENSG00000278869   10.08005   10.10406   10.90964   9.769367   10.19157
#> CABP7-DT          10.08005   10.10406   10.90964   9.769367   10.19157
#>                 ACTTCGTTGA ACTTCTCCTG ACTTGAGGAA ACTTGTAAGG AGAACCACGG
#> NAMPT             13.99649   12.88988  13.439895   12.75603   14.19658
#> ENSG00000278869   10.08960   10.56795   9.980464   10.43410   10.49614
#> CABP7-DT          10.08960   10.56795   9.980464   10.43410   10.49614
#>                 AGAAGCAATC AGACCGTTAT AGACTAGCAT AGAGATGCAG AGAGCTTACA
#> NAMPT             14.22715  14.205986   13.93058   13.96693   14.25161
#> ENSG00000278869   10.13969   9.958058   10.76066   10.26649   10.00369
#> CABP7-DT          10.13969   9.958058   10.76066   10.26649   10.00369
#>                 AGAGTGTAAC AGAGTTCTGC AGATAGTGCT AGCAATGCGC AGCATGTCAT
#> NAMPT             13.66390   14.53844  13.406628   13.95211   14.20371
#> ENSG00000278869   10.20447   10.63155   9.947196   10.04522   10.29682
#> CABP7-DT          10.20447   10.63155   9.947196   10.04522   10.29682
#>                 AGCCACTAGC AGCCAGAATA AGCCAGCTCT AGCGATAACG AGCTATTCCA
#> NAMPT             14.39311  14.045900  13.896092   13.11280  14.355971
#> ENSG00000278869   10.14518   9.958437   9.989202   10.30545   9.963654
#> CABP7-DT          10.14518   9.958437   9.989202   10.30545   9.963654
#>                 AGGCGTCTGT AGGTCACCAA AGTAACTCAC AGTAAGCGTT AGTCCACGTA
#> NAMPT             14.59436   13.67424   13.90536   14.07932  14.083966
#> ENSG00000278869   10.20205   10.21480   10.20492   10.37888  12.318431
#> CABP7-DT          10.20205   10.21480   10.20492   10.37888   9.996503
#>                 AGTCTCGGCA AGTCTGTACG AGTGCAATGT ATAACGCCTC ATAAGAGGTC
#> NAMPT             14.48354   13.37684   13.81293   14.11309   14.19242
#> ENSG00000278869   10.39608   10.20691   10.35350    9.86516   10.49198
#> CABP7-DT          10.39608   10.20691   10.35350    9.86516   10.49198
#>                 ATAAGGTGCA ATACACGACA ATACCTCCGG ATAGCAGTGC ATAGGCCATT
#> NAMPT             13.59311   14.15351  13.380159   12.64341   13.91116
#> ENSG00000278869   10.13368   10.24661   9.920727   10.32148   10.00427
#> CABP7-DT          10.13368   10.24661   9.920727   10.32148   10.00427
#>                 ATATCCGCAT ATCAGCACTT ATCAGCGAGG ATCCAATACG ATCCGCTGTG
#> NAMPT             13.20469   14.30919   14.28599  13.835359   13.51872
#> ENSG00000278869   10.39733   10.40230   10.37910   9.928468   10.34879
#> CABP7-DT          10.39733   10.40230   10.37910   9.928468   10.34879
#>                 ATCCGTCCAT ATCGACGGCT ATCGCGATTA ATCGGTAGGC ATCTAAGGAG
#> NAMPT            13.655788   14.27063   13.72080   13.77757  14.418766
#> ENSG00000278869   9.955348   10.36374   10.02036   10.07713   9.895204
#> CABP7-DT          9.955348   10.36374   10.02036   10.07713   9.895204
#>                 ATGACGGTAA ATGACTCAGT ATGCGGACTG ATGGTCTTAG ATGTGGAAGC
#> NAMPT             14.47485   13.83903  14.708160   13.34782   13.85146
#> ENSG00000278869   10.22693    9.93214   9.850179   10.54047   10.39203
#> CABP7-DT          10.22693    9.93214   9.850179   10.54047   10.39203
#>                 ATTATCGGAC ATTCGGAACA CAACAATCCA CAAGAAGCAT CAAGATGAGG
#> NAMPT             14.38714   12.83998   14.14760   13.75141  13.759497
#> ENSG00000278869   10.29968   10.51805   10.24071   10.29198   9.852606
#> CABP7-DT          10.29968   10.51805   10.24071   10.29198   9.852606
#>                 CAAGCCAACG CAAGTGGATC CACAGTTCAT CACGAGTCTG CACGCTCCAA
#> NAMPT             12.79278   14.13392   14.03263   14.62014   14.13498
#> ENSG00000278869   11.20782    9.88599   10.33219   10.09658   10.43454
#> CABP7-DT          11.20782    9.88599   10.33219   10.09658   10.43454
#>                 CACTGAGCAC CAGATCAATG CAGTGCTCTT CAGTTAAGCA CATAGCTATC
#> NAMPT             14.35395   13.84167   14.23031   13.81527   13.89587
#> ENSG00000278869   10.10602   10.38224    9.98238   10.35584   10.19543
#> CABP7-DT          10.10602   10.38224    9.98238   10.35584   10.19543
#>                 CATCACCACC CATGTACGCC CATTACACTG CATTCGACGA CCAACTATGG
#> NAMPT             14.75381   13.92909   14.08192  13.895667   12.88798
#> ENSG00000278869   10.50588   10.02220   10.38148   9.988777   10.56605
#> CABP7-DT          10.50588   10.02220   10.38148   9.988777   10.56605
#>                 CCAAGGAGTT CCAATTGTTC CCACAAGTGC CCAGCTTAGT CCATAACTTG
#> NAMPT            14.453685   14.38994  14.464874   14.09772  13.624526
#> ENSG00000278869   9.809829   10.30248   9.941312   10.92780   9.717635
#> CABP7-DT          9.809829   10.30248   9.941312   10.92780   9.717635
#>                 CCATAGATCA CCGAACAAGC CCGAAGACCT CCGAATAGTG CCGACTTCTC
#> NAMPT            14.192856   13.64834   14.32621   14.43232   14.71488
#> ENSG00000278869   9.800538   12.06338   10.62577   10.04000   10.32257
#> CABP7-DT          9.800538   10.47842   10.62577   10.04000   10.32257
#>                 CCGATCCACT CCGATGATAC CCGCGTTATG CCGCTAGCTT CCGGAGTATC
#> NAMPT             13.59321   14.04795   13.86349   14.13613   14.81530
#> ENSG00000278869   12.00824   10.58852   10.16305   10.22924   10.06041
#> CABP7-DT          10.42328   10.58852   10.16305   10.22924   10.06041
#>                 CCGGCCAATT CCGGTCTCTA CCGTACGATG CCGTCAGAAC CCTAGACACG
#> NAMPT             13.99348   14.58924   13.39697  14.166851   13.85250
#> ENSG00000278869   10.29304   10.06568   10.58962   9.774533   10.15206
#> CABP7-DT          10.29304   10.06568   10.58962   9.774533   10.15206
#>                 CCTAGTTGAG CCTATTCTGT CCTCAACCGA CCTCCATAAG CCTGATGCCA
#> NAMPT             14.44526   14.80778   14.62671   14.11069   13.94918
#> ENSG00000278869   10.19734   10.55986   10.23439   10.20380    9.86172
#> CABP7-DT          10.19734   10.55986   10.23439   10.20380    9.86172
#>                 CCTGCAATAC CCTTGTATTC CGAGATCTCT CGAGGAACAA CGATAACCGC
#> NAMPT             13.49066   13.96961   13.38090   14.73580   14.19375
#> ENSG00000278869   10.32073   10.06272   10.57355   11.67691   10.28686
#> CABP7-DT          10.32073   10.06272   10.57355   10.09195   10.28686
#>                 CGATCCTGTG CGCCAACCAT CGCCAGTGTT CGCGGATTCA CGCTTCTTGG
#> NAMPT            14.069789   13.76564   14.45898  13.696850   14.54643
#> ENSG00000278869   9.982326   11.44371   10.06667   9.996411   10.29851
#> CABP7-DT          9.982326   11.44371   10.06667   9.996411   10.29851
#>                 CGGAATACAC CGGAGATTGG CGGAGCTCAA CGGATCGGTA CGGATTCTAG
#> NAMPT             14.10723   14.45082   13.96841  14.499409   13.11294
#> ENSG00000278869   10.40679   10.36335   10.26797   9.855552   10.30559
#> CABP7-DT          10.40679   10.36335   10.26797   9.855552   10.30559
#>                 CGGCAACTTA CGGCTCATCA CGGTCGTATT CGGTGACATC CGTAACGGAT
#> NAMPT             12.82179   14.11113   14.16814   12.88258   14.34739
#> ENSG00000278869   10.01443   10.41069   10.08067   10.56065   10.09946
#> CABP7-DT          10.01443   10.41069   10.08067   10.56065   10.09946
#>                 CGTAAGATTC CGTACTGTAA CGTAGAAGAC CGTCCTAGGA CGTCGGCAAT
#> NAMPT             13.64759   13.57998   14.45721  13.888355   13.45407
#> ENSG00000278869   10.18816   10.12055   10.36975   9.981465   10.28415
#> CABP7-DT          10.18816   10.12055   10.36975   9.981465   10.28415
#>                 CGTGAGTTAT CGTGTCAAGC CTAACTTCAG CTAATAGCGT CTACACCAGG
#> NAMPT             14.21134   12.86255  14.389230   13.88852  14.203103
#> ENSG00000278869   10.12387   10.54062   9.996913   12.01405   9.955175
#> CABP7-DT          10.12387   10.54062   9.996913   10.42908   9.955175
#>                 CTAGCACAAT CTATGAACGG CTCAAGGACC CTCACCTGTC CTCCTATTGT
#> NAMPT            13.973762  14.283509   14.02354   14.41938  13.524988
#> ENSG00000278869   9.886299   9.891191   10.32310   10.02706   9.824548
#> CABP7-DT          9.886299   9.891191   10.32310   10.02706   9.824548
#>                 CTCGCAACGT CTGGATTGAC CTGTCTGTGT CTTATGTTGC CTTCATATCG
#> NAMPT            13.903956   14.25770   13.56746   14.12393  13.892189
#> ENSG00000278869   9.997065   10.17024   10.39753   10.21704   9.804727
#> CABP7-DT          9.997065   10.17024   10.39753   10.21704   9.804727
#>                 CTTGCTGACG CTTGGATCTT GAAGAGTTCT GAAGGATTAG GAATCGAGCC
#> NAMPT             13.53461   14.10169   14.60146   13.85138   13.17640
#> ENSG00000278869   10.72725   10.01423   10.20914   10.15094   10.36904
#> CABP7-DT          10.72725   10.01423   10.20914   10.15094   10.36904
#>                 GAATCTTCTC GAATTACGGC GACCATCTAA GACGACCACA GAGAACGAAG
#> NAMPT             13.20838   14.26026  14.169591   14.15530   14.32893
#> ENSG00000278869   10.03845   10.35336   9.921664   10.45486   11.66597
#> CABP7-DT          10.03845   10.35336   9.921664   10.45486   10.08101
#>                 GAGACAAGGC GAGACATCTT GAGCGAGTCA GAGTAGACCA GATACGCTTA
#> NAMPT             14.17458   13.31867   12.82448   13.94739   12.69029
#> ENSG00000278869   11.85265   10.14874   10.50256   10.04050   11.10533
#> CABP7-DT          10.26769   10.14874   10.50256   10.04050   11.10533
#>                 GATAGACTGT GATAGAGGCG GATAGGTCAA GATATCAGGA GATCTCATTC
#> NAMPT             12.95078   13.50194   14.17048   13.76605   13.58436
#> ENSG00000278869   10.62885   10.69458   10.08301   10.30662   10.12493
#> CABP7-DT          10.62885   10.69458   10.08301   10.30662   10.12493
#>                 GATCTGGTCG GATGAGTGAC GATGGATACA GATGTGACAG GATTGCACGC
#> NAMPT             13.44772  14.225836   14.69629   13.99576   10.88166
#> ENSG00000278869   10.27780   9.977908   10.17273   10.08887   10.88166
#> CABP7-DT          10.27780   9.977908   10.17273   10.08887   10.88166
#>                 GCACACTATA GCACTGCGTT GCACTTAATC GCAGGAGATG GCAGTACTGG
#> NAMPT             14.06808   13.50116  13.875953   13.89685  14.519691
#> ENSG00000278869   10.36764   10.33124   9.969063   10.43742   9.996129
#> CABP7-DT          10.36764   10.33124   9.969063   10.43742   9.996129
#>                 GCATATGAGT GCATCCGATC GCCAAGTACA GCCACGATTC GCCATAGGTT
#> NAMPT             14.49169   14.25239   13.43202   14.41545   13.66035
#> ENSG00000278869   10.24377   10.00446   10.26210   10.02313   10.20092
#> CABP7-DT          10.24377   10.00446   10.26210   10.02313   10.20092
#>                 GCCATATCGA GCCGTCAATA GCCTGGACAT GCGTAATTAC GCTATTATCC
#> NAMPT            14.140329   13.76553  14.089849   13.80946  13.486204
#> ENSG00000278869   9.892401   10.30610  11.426884   10.35003   9.785764
#> CABP7-DT          9.892401   10.30610   9.841921   10.35003   9.785764
#>                 GCTCAGTAAT GCTGCTTATA GGAATAAGCA GGACGATGCT GGCATCGTGA
#> NAMPT             13.63412   13.83237   14.28263   12.39894   13.61543
#> ENSG00000278869   10.82677   10.13193   10.37574   10.07701   10.44550
#> CABP7-DT          10.82677   10.13193   10.37574   10.07701   10.44550
#>                 GGCATTATTG GGCCGAGATT GGCGCTATAA GGCGTTAAGT GGCTATTGAT
#> NAMPT             14.04289   13.62700  14.160058   13.67315   13.21275
#> ENSG00000278869   10.13600   10.16757  11.497093   10.21372   10.04282
#> CABP7-DT          10.13600   10.16757   9.912131   10.21372   10.04282
#>                 GGCTGCTACT GGTAATGTGT GGTGTTCACC GGTTCACTGG GTAACCAGTA
#> NAMPT             13.78544   14.27067   14.54249   14.02450   13.62434
#> ENSG00000278869   10.32600   10.02274   10.15017   10.56507   10.16491
#> CABP7-DT          10.32600   10.02274   10.15017   10.56507   10.16491
#>                 GTAACCTTGG GTAAGAACCT GTAAGGCTCC GTAATCCACG GTATTGTGGA
#> NAMPT             13.73078   13.77462   13.65875   13.50599   14.16883
#> ENSG00000278869   10.03034   10.07418   10.19932   10.04656   10.08136
#> CABP7-DT          10.03034   10.07418   10.19932   10.04656   10.08136
#>                 GTCCGCATCA GTCCTTCGGT GTCGCTCTCT GTCGGTGACA GTCTCGAGTG
#> NAMPT             14.12045   15.05770   13.97768  13.309835   13.63661
#> ENSG00000278869   10.42001   10.19971   10.27725   9.850403   10.46669
#> CABP7-DT          10.42001   10.19971   10.27725   9.850403   10.46669
#>                 GTCTCTTAAG GTCTTCCGAG GTGACTATAC GTGGTTAATG GTGTGCCTGT
#> NAMPT             14.31048   13.69240  13.944329   14.18121   13.72460
#> ENSG00000278869   10.40359   10.52248   9.856866   10.27432   10.26517
#> CABP7-DT          10.40359   10.52248   9.856866   10.27432   10.26517
#>                 GTGTGTGTCC GTTCATTGCC GTTCCGGTGA GTTCGTCGAA GTTGAATTGG
#> NAMPT             14.26303   14.22101   13.25959   13.86624   13.57186
#> ENSG00000278869   10.17557   10.13355   10.45223    9.95935   10.40194
#> CABP7-DT          10.17557   10.13355   10.45223    9.95935   10.40194
#>                 GTTGATCCGC GTTGTATGCT TAACCGTAGC TAACGTCGAT TAAGGTACGG
#> NAMPT            12.701573  14.043267   14.33462   14.06657  14.598048
#> ENSG00000278869   9.894218   9.955805   10.08669   10.15967   9.843161
#> CABP7-DT          9.894218   9.955805   10.08669   10.15967   9.843161
#>                 TACTACCGCC TAGCGCCAAC TAGGTTGCAA TAGTAGTCTC TAGTCCGCTG
#> NAMPT            14.569266   13.52407   13.96091  14.282132   13.62775
#> ENSG00000278869   9.925409   10.71671   10.26047   9.889814   10.45783
#> CABP7-DT          9.925409   10.71671   10.26047   9.889814   10.45783
#>                 TAGTGGAACT TATCATGCAG TATCGTTACG TCAAGTGCAG TCACAGATAC
#> NAMPT             12.60432   13.34823   13.97041   13.79348   14.55655
#> ENSG00000278869   10.28239   10.17831   10.06351   10.33404   10.03298
#> CABP7-DT          10.28239   10.17831   10.06351   10.33404   10.03298
#>                 TCACCGCCTA TCACGCCACT TCACGTTGGC TCATTGTCCA TCCACACTAG
#> NAMPT             13.97246   13.98971   13.50309   14.64208   13.02373
#> ENSG00000278869   10.27202   10.08282   10.33316   10.11852   10.21638
#> CABP7-DT          10.27202   10.08282   10.33316   10.11852   10.21638
#>                 TCCACGGTCA TCCACTCGCT TCCGACTAAC TCCGTTATCT TCCTAAGAGA
#> NAMPT             14.64278   12.33585   14.71994   14.10845   14.34250
#> ENSG00000278869   10.11922   10.75088   10.19638   10.20156   10.43561
#> CABP7-DT          10.11922   10.75088   10.19638   10.20156   10.43561
#>                 TCCTCTAGTA TCGAAGCATT TCGAGAGAGC TCGCACTTGA TCGCCTACTG
#> NAMPT             13.62554   13.77115   11.95421  14.833341   12.67466
#> ENSG00000278869   10.81819   10.60123   10.36925   9.879145   10.35273
#> CABP7-DT          10.81819   10.60123   10.36925   9.879145   10.35273
#>                 TCGCGTAGCA TCGGCGTTAA TCTACATCCG TCTCTCCTAT TGAAGAAGGT
#> NAMPT             13.52208   14.43634   14.58437   14.92699   14.93352
#> ENSG00000278869   10.35215   10.04402   10.49691    9.97279   10.28966
#> CABP7-DT          10.35215   10.04402   10.49691    9.97279   10.28966
#>                 TGAGCGTTCC TGAGTACGTA TGGAATGGAG TGTCATTCGC TGTGCTTCAG
#> NAMPT             14.47326   13.31961   13.66720   14.58681  14.197272
#> ENSG00000278869   10.08094   10.51225   10.20777   10.49935  11.534307
#> CABP7-DT          10.08094   10.51225   10.20777   10.49935   9.949344
#>                 TGTTCAGGAT TTACACACGT TTACTGTGAC TTATGCCGCG TTCACGGAAG
#> NAMPT             13.63357  14.565143   14.50972   14.45371   14.27686
#> ENSG00000278869   10.17414   9.707162   10.60283   10.06139   10.36997
#> CABP7-DT          10.17414   9.707162   10.60283   10.06139   10.36997
#>                 TTCGAGTGAT TTCTGTACCT TTGGTAACAG
#> NAMPT             14.59300   14.17005  14.297836
#> ENSG00000278869   10.20068   10.71062   9.905519
#> CABP7-DT          10.20068   10.71062   9.905519
#>  [ reached getOption("max.print") -- omitted 497 rows ]